requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-import-export==3.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
import-export==0.3.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
xlrd==2.0.1
xlwt==1.3.0
